+++ tests/conftest.py
@@ -66,3 +66,18 @@ def simple_strategy(
     return DatabaseStrategy(
         [simple_strategy_trunc, simple_strategy_update, simple_strategy_delete]
     )
+
+
+# MONKEYPATCH: assert_not_called_with
+def assert_not_called_with(self, *args, **kwargs):
+    try:
+        self.assert_called_with(*args, **kwargs)
+    except AssertionError:
+        return
+    raise AssertionError(
+        "Expected %s to not have been called."
+        % self._format_mock_call_signature(args, kwargs)
+    )
+
+
+Mock.assert_not_called_with = assert_not_called_with